+++ oi_cli2/core/DI.py
@@ -1,6 +1,7 @@
 DI_ACCMAN = 'am'
 DI_TEMPMAN = 'tm'
 DI_DB = 'db'
+DI_DB_COOKIES = 'db_cookies'
 DI_LOGGER = 'logger'
 DI_CFG = 'config_folder'
 DI_HTTP = 'http'