+++ oi_cli2/utils/Logger.py
@@ -34,8 +34,8 @@ def getLogger(logger_path):
     # remove default handler
     logging.getLogger().handlers.clear()
 
-    fileformatter = logging.Formatter('[%(asctime)s %(levelname)s %(filename)s %(funcName)s %(lineno)d]%(message)s')
-    streamformatter = logging.Formatter('%(process)s: %(filename)s %(lineno)d %(levelname)-8s %(message)s')
+    fileformatter = logging.Formatter('[%(asctime)s %(levelname)s %(filename)s %(funcName)s %(lineno)d]: %(message)s')
+    streamformatter = logging.Formatter('[%(levelname)s]: %(message)s')
 
     # file
     fh = logging.FileHandler(logger_path)