+++ oi_cli2/utils/db.py
@@ -1,6 +1,7 @@
 import json
 import logging
 import os
+from typing import Optional
 
 
 # write through , read through
@@ -21,7 +22,7 @@ class JsonFileDB:
     with open(self.file_path, "r") as json_file:
       try:
         data = json.load(json_file)
-      except Exception as e:
+      except Exception:
         data = {}
 
     with open(self.file_path, "w") as json_file:
@@ -29,14 +30,16 @@ class JsonFileDB:
       json.dump(data, json_file, indent=2)
     return True
 
-  def load(self, col_name: str) -> any:
+  def load(self, col_name: str) -> Optional[any]:
     if not os.path.exists(self.file_path):
       os.mknod(self.file_path)
 
     with open(self.file_path, "r") as json_file:
       try:
         data = json.load(json_file)
-      except Exception as e:
+      except json.JSONDecodeError:  # empty file or wrong file, type(e).__name__
         data = {}
+      except Exception as e:
+        raise e
 
       return data.get(col_name)